requirements.txt
Django>=5.0,<5.1
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
whitenoise
django-allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe
python-dotenv
python-environ
django-compressor
cryptography
django-extensions
Werkzeug
pyOpenSSL